api/requirements-dev.txt
pytest==8.3.2
vcrpy==6.0.1